V220: fp16→fp32 LUT migration — +27 dB on flat patterns, +21 dB on two_value_split
Replace all fp16 usage in the active encode path:
- d_lut_in (__half*) removed from struct; only d_lut_in_f32 (float*) remains
- encode_ebcot was incorrectly passing fp16 LUT to kernel_rgb48_to_xyz12
- d_a[3]/d_b[3] (__half*) removed from struct; fp32 buffers only
- kernel_rgb48_to_xyz12 updated to const float* LUT (no __half2float overhead)
- gpu_ebcot.h default DWT_T changed from __half to float
- Dead fp16 pipeline functions wrapped in #if 0 (launch_comp_pipeline,
  rebuild_comp_graphs, rebuild_v42_comp_graphs, run_dwt_and_build_codestream)

PSNR improvements (vs V219 baseline):
  flat_30000:      72.2 → 99.0 dB (+26.8 dB, lossless)
  two_value_split: 72.1 → 92.8 dB (+20.7 dB)
  single_impulse:  71.6 → 80.4 dB (+8.8 dB)
  h_gradient:      72.8 → 77.8 dB (+5.0 dB)
  v_gradient:          → 77.7 dB
  h_bars_8:        72.7 → 76.1 dB (+3.4 dB)
  v_bars_8:        72.9 → 76.4 dB (+3.5 dB)
  ramp_small_range:72.1 → 77.3 dB (+5.2 dB)
  photo_synth:     58.1 → 59.1 dB (+1.0 dB)
  checker_64:      28.2 → 28.3 dB (structural issue, unchanged)
verify_roundtrip: 195/195 pass, 0 fail

+++ src/lib/gpu_ebcot.h
@@ -651,13 +651,13 @@ struct CodeBlockInfo {
  * For __half storage we go via __half2float; for float storage we just use __ldg.
  * Marked __forceinline__ so the dispatch is purely compile-time. */
 template<typename DWT_T>
-__device__ __forceinline__ float dwt_load(const DWT_T* p) { return __half2float(__ldg(p)); }
+__device__ __forceinline__ float dwt_load(const DWT_T* p) { return __half2float(__ldg(p)); }  /* fp16 fallback (unused) */
 template<>
 __device__ __forceinline__ float dwt_load<float>(const float* p) { return __ldg(p); }
 
 /* V196: kernel_ebcot_t1 templated on DWT_T to support both FP16 (existing fast
  * path) and FP32 (new high-precision correct path) DWT input. */
-template<bool FAST4, int MAX_BP = (FAST4 ? 4 : 10), typename DWT_T = __half>
+template<bool FAST4, int MAX_BP = (FAST4 ? 4 : 10), typename DWT_T = float>
 __global__ __launch_bounds__(64, 16) void kernel_ebcot_t1(
     const DWT_T* __restrict__ d_dwt,    /* DWT output coefficients (d_a[c]) */
     int dwt_stride,                      /* row stride of DWT array (= image width) */
